Rename class GspellAutomaticSpellChecker -> GspellInlineCheckerGtv
I prefer "Inline" instead of "Automatic". With Automatic, one could
think that the misspelled words are automatically fixed, which some word
processors do e.g. to add an uppercase letter at the beginning of a
sentence or other kind of annoying features like that.

"InlineChecker" has a better contrast with "CheckerDialog".
"Gtv" is for GtkTextView, so in the future there is a place for
GtkEntry.