Change API version to 1
3.0 was chosen because it's the current GTK+ version, so it would have
been easier to remember "gspell-3.0" goes with "gtk+-3.0".

But let's do our own versioning. It's strange to start with "3.0". And I
don't know yet if gspell will be tightly bound to the GTK+ API versions,
maybe for GTK+ 4 there will be no need to break the gspell API so gspell
can keep the 1 version.

Also, use only the major version (1), not 1.0. It's sufficient and more
logical. There won't be an API break for, say, 1.2. API breaks are only
for new major versions.