language-chooser-dialog: add a GtkDialogFlags parameter
To be able to use an headerbar and have more options.
The app must have the choice to use an headerbar when using directly the
dialog, because for some cases it is not suitable. For example in gedit,
a Help button is added, but with an headerbar it is not ideal.

For the LanguageChooserButton, use an headerbar for the dialog, since it
is not possible to add more buttons in that case and an headerbar looks
nicer in my opinion.