checker: rename signals
There was a GI warning because the methods add_word_*() are not the same
as the signals. For the signals, it is not user-friendly to have the
word_length paramaters, it's more convenient to have the string already
nul-terminated. But for the methods it's better to have the word_length
parameters. Passing -1 is not complicated, but allocating a new string
is not convenient.

The signal names are also better like that, because when the signal is
emitted, the action is already done. The action cannot be stopped. It's
just to inform about the event. Otherwise the action would need to be
executed in the object method handler for the signal.