Relicense gspell from GPLv2+ to LGPLv2.1+
Every contributor, including myself, agree to re-license the code to
LGPL.

Evan Martin, the original author of GtkSpell, has sent me a private
e-mail to say that he agrees.

The GNU project recommends to use the GPL license for libraries that
implement something innovative. But spell checking is implemented in
almost every development platform, free software or proprietary. So the
LGPL permits to have wider adoption in that case. See:
http://www.gnu.org/licenses/why-not-lgpl.html

https://bugzilla.gnome.org/show_bug.cgi?id=758217