inline-checker-gtv: adjust subregion only when needed
Most of the time, it is not needed to adjust the subregion before adding
it to the scan_region. The subregion will be correctly adjusted in
check_subregion().

The exception is for the delete-range handler, which is connected with
the after flag, so @start == @end, so the subregion is empty and needs
to be adjusted to include the neighbor words that are potentially
modified.