inline-checker: do not have reference cycles
Although GtkTextView is a widget, gtk_widget_destroy() is sometimes not
called in applications because e.g. gtk_notebook_remove_page() is called
which does a simple g_object_unref(). Even if it is a best-practice to
call gtk_widget_destroy() when a widget is no longer needed, it is not
guaranteed that it is the case. So it's better to avoid ref cycles in
the first place, when easily feasible, which is the case here.

Not having ref cycles was actually a reason to not expose
GspellInlineCheckerText. Otherwise weak references would have been
needed, which complicates a little the code since it needs to handle the
NULL case.