Rename GspellNavigatorGtv -> GspellNavigatorText
For GtkTextView, the suffix used in gspell is now "Text". In the future
the support of GtkEntry will have the suffix "Entry". Text and Entry fit
well together, whereas Gtv and Entry is awkward.

The purpose of the suffix "Gtv" was meant to be able to distinguish it
clearly from other text widgets from other libraries like WebKitGTK+.
But gspell itself will not add support for other libraries, it should
only depend on GTK+.