inline-checker-text-buffer: remove debug code
It gets in the way, and anyway when debugging something we are usually
interested by different information each time.

To be sure that the code is correct, I prefer code clarity, to make it
easy to read and understand. By adding lots of debug code, we make the
code less easy to read, and thus there can be more bugs, which is
exactly the thing that we want to avoid…