inline-checker-text-buffer: fix constant redrawing
When check_current_word was FALSE, the GtkTextView was constantly
redrawn. For two reasons:
- because gtk_text_buffer_remove_tag() was called on the current word,
  even if the tag was not present. Which always queued a redraw.
- because of the queue_draw() hack, which was also always called when
  check_current_word is FALSE.

So, for the current word handling, do the queue_draw() hack only if the
tag is removed. And remove the tag only if the tag is present.

Also, queue a redraw on all views, because if two views display the same
region, only one view was redrawn since the scan_region is already empty
for the second view.

Thanks to Sebastian Keller for the bug report.

https://bugzilla.gnome.org/show_bug.cgi?id=768306