Add GspellEntryBuffer class
Exactly the same as GspellTextBuffer, to have a consistent API.

I think most of the time GtkEntry is used alone, app developers don't
care about GtkEntryBuffer. But it is easy to get the GtkEntryBuffer of a
GtkEntry, and for those who use GtkEntryBuffer, gspell supports it.
There will anyway be a basic_setup() function in GspellEntry.