Add GspellEntry API
Similar to GspellTextView, but without the enable-language-menu property
(it'll be always TRUE). For GtkEntry, inline spell checking is the only
feature to do, so a context menu is always needed. For GtkTextView,
there can be a checker dialog or a checker bar, so the context menu is
not always wanted there.

A basic_setup() function will be added later.