EntryUtils: get list of words without preedit string
Don't take into account the preedit string for spell checking.  It is
anyway planned to *not* spell check the word being typed (so the word at
the cursor, with the preedit string). So it actually doesn't change the
outside result if we take into account the preedit string or not.