build: better order of toplevel SUBDIRS
- Regroup source code.
- Have docs/ at the end because gtk-doc is slow.

When running `make`, I can look at the output when it's building the
source code, to see if there is any warning. But for the docs I prefer
to do something else in the meantime, then come back later to see if
there is any warning (so with the docs at the end, I don't need to
scroll up).

(Well, currently building the gspell docs doesn't take too much time,
but for GtkSourceView for example it is the case, and putting the docs
at the end makes the build consistent with GtkSourceView).