InlineCheckerTextBuffer: replace g_assert_cmpint() by g_return_if_fail()
To avoid a crash. I don't understand why it sometimes fails, probably
because of a bug in a lower-level library (GTK+ or Pango). We don't have
steps to reproduce the problem, just a backtrace on the Fedora Retrace
server.

On master (leading towards the 1.4 version), the g_assert_cmpint() has
not been changed to the g_return_if_fail(), more assertions have been
added instead, so we will likely see again crashes for the 1.4 version,
which will give us more information. For the 1.2 branch I prefer to make
it more stable, while early adopters of 1.3/1.4 can serve as testers.

https://bugzilla.gnome.org/show_bug.cgi?id=778883