InlineCheckerTextBuffer: change underline to PANGO_UNDERLINE_SINGLE
Because PANGO_UNDERLINE_ERROR is buggy:
https://bugzilla.gnome.org/show_bug.cgi?id=763741

It'll permit to remove a hack in gspell that tries to work around the
bug (but doesn't work in all cases).

KWrite (KDE text editor) also inserts the equivalent of
PANGO_UNDERLINE_SINGLE. They maybe have the same bug with wavy
underlines.

The color is defined in gspell-utils.h, because it'll be useful to have
a common place to define the color for both GtkTextView and GtkEntry.