InlineCheckerTextBuffer: reduce timeout duration when buffer modified
Now that the check_subregion() function has a faster implementation (in
most cases it takes less than 16.67 ms to execute on my computer, for a
single line of text), it's useless to wait a long time before checking
the spelling.

With the long line provided in:
https://bugzilla.gnome.org/show_bug.cgi?id=776811
writing new text or deleting text is now responsive.

60 FPS is not always achieved, but it feel responsive.