gspell-app: suffix gspell API version to the binary name
Ideally there should be only one gspell-app binary that uses the latest
version of the gspell library. But to achieve that, I think the only way
is to create another git repository for gspell-app. But I don't want to
duplicate all the build system boilerplate and the resulting maintenance
of having another git repo to take care of. So I prefer to keep
gspell-app in the gspell repo.

When gspell 2 will be released, its gspell-app must not conflict with
the gspell 1 one. So this commit adds the API version (which is also the
major package version) as a suffix: gspell-app1, gspell-app2, etc.

gspell-app1, gspell-app2, etc will also permit to test different
versions of gspell (useful when reporting a bug), although the UI tests
in tests/ are more complete (but not installed).