Checker: expose EnchantDict
I initially didn't want to expose the use of Enchant, because I thought
that gspell could use directly hunspell instead. But I think this won't
happen, Enchant provides a C API while hunspell is in C++, so it's
simpler to use Enchant in gspell.

So the use of Enchant is now part of the API, if it needs to change in
the future, we can just bump the API/major version of gspell. This will
anyway happen regularly to port gspell to new major versions of GTK+.

This commit will normally permit to implement the following feature in
gnome-builder (even if I would prefer that it's implemented in gspell,
to make the feature available to all applications):
https://bugzilla.gnome.org/show_bug.cgi?id=765616