Use Enchant 2
Enchant has a new maintainer and the project is again under development.

There has been some API/ABI breaks in Enchant 2, but they don't affect
gspell.

Enchant 2 >= 2.1.3 is now parallel-installable with Enchant 1, so it'll
normally not cause problems for Linux distribution packagers to ship
Enchant 2 and the next version of gspell. See:
https://github.com/AbiWord/enchant/issues/162

Strictly speaking, this is an API break for gspell as well, since there
is the gspell_checker_get_enchant_dict() function that exposes Enchant
in the public API of gspell. This function has been added in the
previous development cycle for gnome-builder needs, but gnome-builder
doesn't use yet this function. I've asked on IRC to the gnome-builder
developers if it's fine to switch to Enchant 2, and they said that it is
not a problem. So I prefer to follow a pragmatic approach, in practice
switching gspell to Enchant 2 will normally not cause problems for
applications, and that way the apps will automatically benefit from a
newer version of Enchant (especially gedit which is currently
unmaintained, so it's unlikely that gedit would have switched to
gspell-2).

https://bugzilla.gnome.org/show_bug.cgi?id=790994