README: some improvements
The 0.1 version is quite old, at this point I think nobody will use it,
so I remove the warning about it.

And since the GTK+ requirement is still 3.20, it's possible to install
gspell 1.8 or 1.6 on older versions of distributions. It would be
beneficial, because gspell >= 1.6 works better than the previous
versions.