gspell-language.c: Support ISO639 2T locale codes on Windows
GetLocaleInfoEx() does not enable us to look for locales that are in the form
of <iso639_2t_code>_<iso3166_code>, but does allow us to enumerate all the
available locales supported by the running system and find the translated
language names, ISO 639 2T locale code and ISO3166 country code for each of the
locales.

Note that the language names retrieved by GetLocaleInfoEx() contains the
country/region name(s) already, if they exist for the given locale code, so
there is no need to separately look for the country name that is denoted by
the ISO3166 code portion.

So, we still need to make use of the ISO639 HashTable to cache up the ISO 639
2T+ISO 3166 country code combination, along with their corresponding translated
language/locale names