Move ISO Code handling parts to a new source file
Move the parts in gspell-language.c regarding retrieving the (translated)
language names from either the ISO-codes files or the Windows APIs into a file
of their own, so that it is easier to re-use these items for other uses.

Also do a minor fix-up when handling queries done with ISO639/2T language codes
on Windows.