build: fix `make distcheck` with newer gtk-doc version
There was the following error when running `make distcheck`:

ERROR: files left in build directory after distclean:
./docs/reference/gspell-1.0.actions

This should most probably be fixed in gtk-doc itself, for the autotools
files that it provides. In the meantime, be able to release gspell.

It also adds the file to the .gitignore via git.mk.