doap: small update/simplification
gspell is not part of GNOME Core, but it's a core dependency.
In case of doubt, it's better to remove the <category>, and I don't
think it's super important after the switch from cgit to GitLab (in cgit
the git repos were grouped into different categories).

And, the less the better. Less things to keep up-to-date afterwards.

Space out a bit the XML elements, for an easier reading.