Remove gspell-version.h, replace by direct use of G_MODULE_EXPORT
The less the better.

The GSPELL_AVAILABLE_IN_* macros were intended to be used, perhaps one
day, to have something like GTK has: to check at compile time that we
don't depend on too recently added symbols, if we need to deploy a
program on a distro with an older version of the library. Or, use
conditionally newer symbols, etc. With deprecation macros and
everything.

But I prefer not to have all the infrastructure for that in a small
library like gspell. If creating a small library involves adding lots
and lots of scafffolding, it's too much work compared to the size of the
library.