meson: bump soversion
gspell_get_resource() was exported by mistake with the Autotools build
system. That function is no longer exported with Meson.

This is the only difference catched with:
$ nm -D libgspell-1.so.2.3.2 | grep T

But this requires to bump the soversion.

Set the soversion to a single number: 3. The soversion with three number
components was set to have the same as used with Libtool. But the
Libtool versioning is a bit complicated, so from now on, keep thing
simpler by having the soversion as a single number. This should be
sufficient for all OS distributions. If not, please open a bug report.