meson: install tests
Set `Type=session-exclusive` in all the *.test files, because I'm not
sure that the tests can run in parallel (if they add words to
dictionaries or things like that). In case of doubt, it's better to set
a setting that is known to work.

Fixes https://gitlab.gnome.org/GNOME/gspell/-/issues/37